PropBot/requirements.txt
openai==0.27.0
langchain==0.0.123
pinecone-client==2.2.1
streamlit==1.21.0
PyPDF2==3.0.1
python-dotenv==1.0.0